requirements.txt
Django
django-crispy-forms
django-storages[google]
gcsfs
django-cleanup